Xcode GM supported (#53)
* Updated for Swift 3

* Fixed tests

* Add support exporting nested JSON objects

* test for encoding nested JSON objects and remove unneeded imports.

* implementation for top level array and tests for Decoding

* - Solved dynamicType and String initializer

* - Fixed syntax upgrade issues

* - Fixed unit tests

* - Updated the Array extension to adopt swift 3 new syntax

* - Fixed the Mirror doesn't have displayStyle for Swift default type, like String, Int, etc

* - Updated JSONCompatible toJSON bug

* Xcode GM supported

* Upgraded version

+++ JSONCodableTests/NestItem.swift
@@ -6,11 +6,10 @@
 //
 //
 
-import Foundation
 import JSONCodable
 
 struct NestItem {
-    let areas: [[Float]]
+    let areas: [[Double]]
     var places: [[String]]?
     var business: [[Company]]
     var assets: [[ImageAsset]]?
@@ -28,6 +27,4 @@ extension NestItem: JSONDecodable {
             fatalError("\(error)")
         }
     }
-
 }
-
